move $ selector yieldable to addon code, not test
